перезод к типу long double

+++ include/Particle.h
@@ -11,14 +11,14 @@ namespace Particle_Particle_model {
 		vec r, v, F, L, P, a;
 		// TODO: add function
 		Particle():
-			r(0, 0, 0), 
-			v(0, 0, 0), 
-			F(0, 0, 0), 
-			L(0, 0, 0), 
-			P(0, 0, 0),
-			a(0, 0, 0),
+			r(0.0L, 0.0L, 0.0L), 
+			v(0.0L, 0.0L, 0.0L), 
+			F(0.0L, 0.0L, 0.0L), 
+			L(0.0L, 0.0L, 0.0L), 
+			P(0.0L, 0.0L, 0.0L),
+			a(0.0L, 0.0L, 0.0L),
 			m(M/N),
-			E(0)
+			E(0.0L)
 		{}
 		void Evaluate_a() { a = F / m; }
 
@@ -27,14 +27,14 @@ namespace Particle_Particle_model {
 	inline vec F(const Particle& p1, const Particle& p2)
 	{
 		const vec r_ij = p2.r - p1.r;
-		const long double r =sqrt(r_ij.module_2()+r_c);
+		const long double r =sqrtl(r_ij.module_2()+r_c);
 		return  r_ij*(G * (p2.m*p1.m) / (r*r*r));
 	}
 
 	inline vec calc_f(const Particle& p1, const Particle& p2)
 	{
 		const vec r_ij = p2.r - p1.r;
-		const long double r = sqrt(r_ij.module_2() + r_c);
+		const long double r = sqrtl(r_ij.module_2() + r_c);
 		return  r_ij * (G * (p2.m) / (r * r * r));
 	}	
 }
